+++ CHANGELOG.md
@@ -7,13 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+-   Added the who command and more info to help
+-   Only log messages if channel is configured
+
 ## [3.1.3] - 2023-04-10
 
+### Changed
 -   Reduce initial values
 
 ## [3.1.2] - 2023-04-07
 
+### Changed
+
 -   Auto leave young servers
+
+### Added
+
 -   Added admin post to force leave a server
 
 ## [3.1.1] - 2023-04-07